Javascript: Added new command switch: -api <foo> intended to allow users easy selection of the different v8 APIs.
Valid values are currently V8_3_14 or BUILDING_NODE_EXTENSION. Other values will trigger an error, though we may consider changing this if the versions continue to constantly break. The way it works is it just adds a #define <foo> at the top of the generated SWIG file.

This is only valid for v8. Other interpreters should trigger an error. This could change as needed if other interpreters find this useful.

Example:
./preinst-swig -javascript -v8  -api V8_3_14  -o ~/Source/LANICA/SWIG_EXPERIMENTS4/ChipmunkSwigV8Android/jni/chipmunk_wrap_javascript_v8.cpp -DTITANIUM ~/Source/LANICA/ChipmunkSWIG/chipmunk/chipmunk.i