Javascript v8: Bug fix for v8 version selection using -D switch. The macro substitution was wrong because new code trying to be generated using that same macro token was getting substituted which was not the desired outcome. You must use -DUSE_V8_VERSION=<version> instead now.
So we were getting this:
#ifndef 0x031400
#define 0x031400 0x031400
#endif

instead of this:
#ifndef SWIG_V8_VERSION
#define SWIG_V8_VERSION 0x031400
#endif

To avoid this problem, this solution uses a different macro token name, USE_V8_VERSION.
-DUSE_V8_VERSION=0x031400

Also, to make the default case work, the version number needed to be explicitly added to the code generation. Based on the mailing list announcement, I've made that 3.11.10 (0x31101).