Javascript: Initial fix for %rename problem documented here: https://github.com/oliver----/swig-v8/issues/18. It looks like more may be needed requiring an audit of the existing code. But this solves the section for variables.
In exitVariable, if I change the lines that look like:
.replace(T_NAME, state.variable(NAME))
to
.replace(T_NAME, Swig_name_mangle(state.variable(NAME)))
this fixes the problem.

Using state.variable(NAME_MANGLED) does not work though.

This might suggest there are more places that need Swig_name_mangle() to account for the possibility of %rename being in use.

Oliver responded with:
Thanks. You are absolutely right... I have to revisit all the places and consider usage of Swig_name_mangle() where it is possible.